Header hardware targets depend on hardware_claim_headers
A target created by pico_simple_hardware_headers_target must depend on hardware_claim_headers not hardware_claim.

This avoids adding claim.c to the user sources. If the user wanted claim.c to be added to the sources, the user would have used the targed defined by pico_simple_hardware_impl_target.

This is not an issue for hardware_structs since this target does not define sources.

Fixes #957